Update zgen installation command
By default, zgen loads plugins from the master repository. However, when attempting to load the history-search-multi-word plugin, an error is encountered due to its default branch being named "main" instead of "master". To address this issue, you can add the branch name to the command, with the period before the branch name indicating the location within the repository that zgen should search for the plugin. In this case, the period represents the root directory.